models: Add request_timestamp field to ScheduledMessage table.
This just records the time when the user created the
request to schedule a message or set a reminder.

This would be unused but is future-proofing against a world
where we decide to edit the strings or UI to show when you
asked for the thing.

(cherry picked from commit 9d3279caf43e99959941f37504654863ec1b7eb0)